Исправила классы mail
