+++ src/main/java/ru/mail/polis/homework/collections/mail/MailMessage.java
@@ -3,11 +3,12 @@ package ru.mail.polis.homework.collections.mail;
 /**
  * 1 тугрик
  */
-public class MailMessage {
+public class MailMessage extends Mail {
 
     private final String message;
 
-    public MailMessage(String message) {
+    public MailMessage(String sender, String recipient, String message) {
+        super(sender, recipient);
         this.message = message;
     }
