+++ src/main/java/ru/mail/polis/homework/collections/mail/Salary.java
@@ -3,15 +3,15 @@ package ru.mail.polis.homework.collections.mail;
 /**
  * 1 тугрик
  */
-public class Salary extends Sending {
-    private final double salaryAmount;
+public class Salary extends Mail {
+    private final int salaryAmount;
 
-    public Salary(String sender, String recipient, double salary) {
+    public Salary(String sender, String recipient, int salary) {
         super(sender, recipient);
         this.salaryAmount = salary;
     }
 
-    public double getSalaryAmount() {
+    public int getSalaryAmount() {
         return salaryAmount;
     }
 }
